Changed to float to int

+++ main.py
@@ -26,6 +26,6 @@ if string_data:
 if st.button('Send'):
     res = requests.post(API_ENDPOINT2, json={"project_id": question_area, "country" : country , "industry" : inds, "project_type" : proj, "top_number" : top_number})
     df=pd.DataFrame(res.json()["Results"])
-    df["projectid"] = df["projectid"].astype(str).str.replace(".0000", "")
+    df["projectid"] = df["projectid"].astype(str).str.replace(".0000", "").astype(float)
     df["projectid"] = df["projectid"].astype('int64')
     st.dataframe(df)
